requirements.txt
mcp[cli]
google-generativeai
openai
anthropic
requests
requests-oauthlib
pandas
python-dotenv
pydantic
pillow
